Struct google_cloudresourcemanager1_beta1::Project

pub struct Project {
    pub name: Option<String>,
    pub parent: Option<ResourceId>,
    pub project_id: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub project_number: Option<String>,
    pub create_time: Option<String>,
    pub lifecycle_state: Option<String>,
}

A Project is a high-level Google Cloud Platform entity. It is a container for ACLs, APIs, AppEngine Apps, VMs, and other Google Cloud Platform resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

test iam permissions projects (none)

undelete projects (none)

set iam policy projects (none)

get projects (response)

get ancestry projects (none)

update projects (request|response)

get iam policy projects (none)

delete projects (none)

create projects (request|response)

list projects (none)

Fields
name: Option<String>

The user-assigned display name of the Project. It must be 4 to 30 characters. Allowed characters are: lowercase and uppercase letters, numbers, hyphen, single-quote, double-quote, space, and exclamation point. Example: My Project Read-write.

parent: Option<ResourceId>

An optional reference to a parent Resource. The only supported parent type is "organization". Once set, the parent cannot be modified. Read-write.

project_id: Option<String>

The unique, user-assigned ID of the Project. It must be 6 to 30 lowercase letters, digits, or hyphens. It must start with a letter. Trailing hyphens are prohibited. Example: tokyo-rain-123 Read-only after creation.

labels: Option<HashMap<String, String>>

The labels associated with this Project. Label keys must be between 1 and 63 characters long and must conform to the following regular expression: [a-z]([-a-z0-9][a-z0-9])?. Label values must be between 0 and 63 characters long and must conform to the regular expression ([a-z]([-a-z0-9][a-z0-9])?)?. No more than 256 labels can be associated with a given resource. Clients should store labels in a representation such as JSON that does not depend on specific characters being disallowed. Example: "environment" : "dev" Read-write.

project_number: Option<String>

The number uniquely identifying the project. Example: 415104041262 Read-only.

create_time: Option<String>

Creation time. Read-only.

lifecycle_state: Option<String>

The Project lifecycle state. Read-only.